Revert "udfps: Implement default udfps hbm provider"
This reverts commit 363aec1673420b8f3518a7c133800efcb99731e5.

Change-Id: I6bd913aab11d52b6b6d60038d397a6f6952f387f

SystemUI: Use Pixel UDFPS HBM provider
This provides an implementation of the optional UdfpsHbmProvider
interface on Pixel devices.

SystemUI: Add HBM provider for UDFPS on Pixel devices
This implements high brightness mode (HBM) control for the under-display
fingerprint sensor on Pixel devices. It enables and disables local HBM
through the Pixel display HAL, similar to Google's implementation in the
stock OS.

Blocking IPC calls to the display HAL are done on a background thread to
avoid causing frame drops by blocking the main thread, but callbacks are
always run on the main thread as specified by the UdfpsHbmProvider docs.
Care is taken to avoid enabling LHBM when it's already enabled and vice
versa, as doing so causes the display HAL's request to time out and
throw an exception.

For simplicity, global HBM is not supported as it has never been used in
production on Pixel devices.

Change-Id: I4c9f55e3580d2cebc94144cd69227077a104448e